Added flags and logic for animatio, hero-content and resetbutton
Flags to check if a card is open or not, whether camera is still lerping or not.

These flags are used to open the html card delayed after lerping, and to hide the instruction text while a card is open.
This re-appears like usual after the 4 seconds of no interactions.

Removed flickering of text using the reset button, was old no longer needed code.

Also added propagation stops on the html cards to prevent orbitControls from interacting through them.